SAM as teacher, MRL, split code into modules
